requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
pillow
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
